Rewrite channels slides and examples (#88)
